requirements.txt
Django==1.9
psycopg2
djangorestframework
djangorestframework-gis
